+++ runtime/src/commonMain/kotlin/dev/zacsweers/metro/internal/MapFactory.kt
@@ -54,8 +54,8 @@ public class MapFactory<K : Any, V : Any> private constructor(map: Map<K, Provid
     }
   }
 
-  public companion object MapFactoryCompanion {
-    private val EMPTY: Provider<Map<Any, Any>> = InstanceFactory.create(emptyMap())
+  public companion object {
+    private val EMPTY: Provider<Map<Any, Any>> = InstanceFactory(emptyMap())
 
     /** Returns a new [Builder] */
     public fun <K : Any, V : Any> builder(size: Int): Builder<K, V> {